places vc refactor
